requirement.txt
appdirs==1.4.3
astroid==2.4.2
attrs==19.3.0
backcall==0.1.0
baseline==1.1.2
black==19.10b0
bleach==3.1.5
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
easydict==1.9
entrypoints==0.3
flake8==3.7.9
flake8-bugbear==20.1.4
flake8-comprehensions==3.2.2
future==0.18.2
fvcore==0.1.dev200325
glog==0.3.1
gym==0.17.2
h5py==2.10.0
idna==2.9
imageio==2.8.0
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.1
olefile==0.46
opencv-python==4.2.0.34
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
portalocker==1.7.0
prefetch-generator==1.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyglet==1.5.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-gflags==3.1.2
pytorch3d==0.2.0
PyYAML==5.3.1
pyzmq==19.0.0
regex==2020.4.4
requests==2.23.0
scikit-learn==0.23.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
spatial-correlation-sampler==0.2.1
tabulate==0.8.7
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
toml==0.10.0
torch==1.5.0
torchfile==0.1.0
torchvision==0.6.0a0+82fd1c8
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
trimesh==3.6.43
typed-ast==1.4.1
urllib3==1.25.9
visdom==0.1.8.9
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
wrapt==1.12.1
yacs==0.1.6
yapf==0.30.0
